pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "pgcn"
version = "0.1.0"
description = "Plasticity-Guided Connectome Network utilities"
authors = [{name = "PGCN Team"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "pandas",
    "pyarrow",
    "numpy",
    "networkx",
    "tqdm",
    "pydantic",
    "rich",
    "tenacity",
    "caveclient",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "black",
]
models = [
    "torch>=2.2,<3",
]
api = [
    "fastapi>=0.110",
    "uvicorn>=0.23",
]

[project.scripts]
pgcn-cache = "pgcn.connectome_pipeline:main"
pgcn-metrics = "pgcn.metrics:cli"
pgcn-auth = "pgcn.auth:cli"
pgcn-access-check = "pgcn.flywire_access:cli"
pgcn-codex-import = "pgcn.codex_import:cli"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
